add Generator module
tests.pro > helptest.pro > generator > Generator.?pp

still two issues:
- undefined reference when SUBDIRS += helptest \ uncommented
- undefined reference when all Generator function implementations moved
to Generator.cpp
