--- tests/helptest/generator.pro
@@ -1,12 +1,0 @@
-TINO_LIB_DEPENDS += \
-    core
-
-include(../../tests.pri) #not sure
-
-HEADERS += $$PWD\Generator.hpp
-SOURCES += $$PWD\Generator.cpp
-
-INCLUDEPATH += \
-    $$PWD/. \
-    $$PWD/../../src/core/ \
-
