+++ tests/helptest/generator/Generator.cpp
@@ -1,0 +1,2 @@
+#include "Generator.hpp"
+
