+++ tests/helptest/generator/generator.pro
@@ -1,0 +1,8 @@
+TINO_LIB_DEPENDS += \
+    core
+
+include(../../tests.pri)
+
+HEADERS += Generator.hpp
+SOURCES += Generator.cpp
+
