+++ tests/helptest/helptest.pro
@@ -1,12 +1,3 @@
-TINO_LIB_DEPENDS += \
-    core
-
-include(../tests.pri) #not sure
-
-HEADERS += $$PWD\Generator.hpp
-SOURCES += $$PWD\Generator.cpp
-
-INCLUDEPATH += \
-    $$PWD/../../src/core/ \
-    $$PWD/.
-
+TEMPLATE = subdirs
+SUBDIRS += \
+    generator
