+++ tests/tests.pri
@@ -5,4 +5,4 @@ QT += testlib core gui widgets
 CONFIG += qt warn_on console depend_includepath testcase
 CONFIG -= app_bundle
 
-#INCLUDEPATH += ../../helptest/generator/
+INCLUDEPATH += $$PWD/helptest/generator/
