+++ tests/tst_widget/tst_mask/tst_mask.pro
@@ -1,6 +1,7 @@
 TINO_LIB_DEPENDS += \
     core \
-    widget
+    widget \
 
 include(../../tests.pri)
 
+SOURCES += tst_Mask.cpp
